fix the expired PersistentTopic has a closed managedLedger that will throw a ManagedLedgerException (#862)
Fixes #861

### Motivation
See #861 for the reason for the investigation.

Since the namespacebundle has a listener, it can be guaranteed that the topic corresponding to tcm can be removed from the cache after being unloaded and loaded.

But for topic, it does not seem reasonable to create a listener for close, because for the server, there will be a large number of topics .

One more thing, pulsar currently does not have a listener for the close method of a single topic. And there is no need to implement such a listener.

(cherry picked from commit 1565147677632c356ff1ee0353a44122fb642e0f)