[Issue 867]: Implement minimal DeleteRecords (KStreams support) (#871)
KafkaStreams uses the DeleteRecords API to truncate topics to the very end of the topic.

https://github.com/apache/kafka/blob/ba0ebca7a516d4179b6327ddc60b0b49b1265347/streams/src/main/java/org/apache/kafka/streams/processor/internals/TaskManager.java#L441

This patch implements DeleteRecords, using the `truncateAsync` API in Pulsar introduced in 2.8.0.

we cannot make it work for arbitrary offsets, by the way this implementation is still useful, at least for KStreams or for Kafka Users who want to empty a topic.

Please note that Kafka Streams does not break without this feature, because it tries to delete the records in a best effort way.

Fixes #867

(cherry picked from commit 2829e543a5799e725acf4f0adfdced06928f5123)