Fix list offsets failure when ledgers are removed by a rollover operation (#894)
Fixes https://github.com/streamnative/kop/issues/876

There is a corner case that `MessageMetadataUtils.getOffsetOfPosition` fails with `NonRecoverableLedgerException`.

If a managed ledger has multiple ledgers that are all expired. After an automatically rollover controlled by max rollover time), a new empty ledger will be created, then the previous ledgers will be removed. However, a rollover doesn't change the `lastConfirmedEntry`, in `ManagedLedgerImpl#getFirstPosition`, if the first ledger's id is greater than the `lastConfirmedEntry`, the returned position's ledger id will be the same with the `lastConfirmedEntry`'s ledger id. In addition, KoP applies `ManagedLedgerImpl#getNextValidPositionInternal` to the first position. In this method, if the position's ledger id is the same with `lastConfirmedEntry`, it's also treated as valid.

Therefore, in this case, KoP will try to read an entry whose ledger id is not in the managed ledger, and a `NonRecoverableLedgerException` will be thrown. See `KafkaRequestHandler#fetchOffset` for details.

- Handle the `NonRecoverableLedgerException` in `readEntriesFailed` callback. In this case, return the LEO as the offset because we can treat the topic as empty.
- Add `testListOffsetForEmptyRolloverLedger` to `MultiLedgerTest` to reproduce the scenario I've described. After the ledgers are all removed by a rollover, use a Kafka consumer to list offsets. Then produce a message and try to consume it.

Besides, this PR also handles the case that `getFirstValidPosition` returns a position whose ledger is doesn't exist but the managed ledger is not empty. Returns the position of first ledger's entry. This case should never happen except there is a bug in Pulsar.

(cherry picked from commit a83b2ad6e45ee96dd4cdb82c6d6405d0cc337631)