Set default validate method value for OAuth (#1241)
Currently, when we use OAuth in KoP, we must set the `kopOauth2ConfigFile`, but most of the time we only use `token` as the validate method.

We should set the default validate method when `kopOauth2ConfigFile` is not configured.

Set `token` as the default validate method value.

(cherry picked from commit 945132d2e74952823dcaff38bb00f060a0b797cc)