[Authorization] Remove unexpected produce permission validation (#1297)
Fixes https://github.com/streamnative/kop/issues/1295

### Motivation

Currently KoP registers a `InternalProducer` that inherits the broker's
`Producer`  with a unique name for each topic to record the stats so
that REST API. However, when the authoriazion is enabled,
`checkPermissionAsync` will be called on the `InternalProducer` object
when the policies are update in `PersistentTopic#onPoliciesUpdate`.

```java
return cnx.getBrokerService().getAuthorizationService()
        .canProduceAsync(topicName, appId, cnx.getAuthenticationData())
```

The code snippet above is from `Producer#checkPermissionAsync`, we can
see the auth role is the `appId` field (`null` in `InternalProducer`'s
constructor) and the authentication data is from the `cnx` field, which
is the `InternalServerCnx` object in KoP. The role and auth data are
both `null`, if the authorization provider doesn't assume they are not
null, NPE will happen in the `metadata-store` thread

However, before https://github.com/apache/pulsar/pull/13885, a
synchronous method `checkPermission` is called, the NPE will be caught
and `disconnect` will be called after that.

```java
if (cnx.getBrokerService().getAuthorizationService() != null) {
    try {
        if (cnx.getBrokerService().getAuthorizationService().canProduce(topicName, appId,
                cnx.getAuthenticationData())) {
            return;
        }
    } catch (Exception e) {
        // 1. NPE is caught here
        log.warn("[{}] Get unexpected error while autorizing [{}]  {}", appId, topic.getName(), e.getMessage(),
                e);
    }
    log.info("[{}] is not allowed to produce on topic [{}] anymore", appId, topic.getName());
    disconnect(); // 2. The channel will be closed
}
```

So sometimes the `testGrantAndRevokePermission` could fail because the
`KafkaProducer#send` might throw `NetworkException` instead of
`TopicAuthorizationException`.

### Modifications

Override the `checkPermissionAsync` and `checkPermission` methods with
an empty implementation so that this PR can be cherry-picked into
branch-2.8 and branch-2.9.

Then add a test to save the exception stack in
`OauthMockAuthorizationProvider#isSuperUser`. Without this patch, a stack
trace of NPE will be saved in `NULL_ROLE_STACKS`.

(cherry picked from commit 8a95b3d97dad90355969cb48170592f7442f2590)