[Transaction] Return INVALID_PRODUCER_EPOCH error when the transaction produces epoch fence happened (#1443)
Fixes: #1424

### Motivation

When the KoP handled the end transaction request, it returned the `UNKNOWN_SERVER_ERROR` error.

I can't reproduce this in Flink Jobs, but when I review the end transaction code, it is possible to be caused by the produce epoch fence since we are using the `UNKNOWN_SERVER_ERROR` error as the produce epoch fence error.

However, in Kafka, when the produce epoch fence, it will return an `INVALID_PRODUCER_EPOCH` error, we should keep it the same.

Error log:
```
2022-08-05T08:37:26,952+0000 [pulsar-ph-kafka-64-30] ERROR io.streamnative.pulsar.handlers.kop.coordinator.transaction.TransactionCoordinator - Aborting append of COMMIT to transaction log with coordinator and returning UNKNOWN_SERVER_ERROR error to client for Source: Custom Source -> Sink: Unnamed-7df19f87deec5680128845fd9a6ca18d-4's EndTransaction request
```

### Modifications

* Replace `UNKNOWN_SERVER_ERROR` error to `INVALID_PRODUCER_EPOCH`

(cherry picked from commit 6fde976590e5a346e855c822211a08858658a4e7)