[fix] Use correct path for pulsar client TlsTrustCertsFilePath (#1604)
### Motivation

This is a trivial fix. The `getTlsCertificateFilePath` is meant to be
the tls certificate chain when sending to a peer, and the
`getTlsTrustCertsFilePath` is meant to be the root CAs.

This PR could technically break installations. I'm unsure how we'd like
to handle backwards compatibility here.

Note that this PR is the same as
https://github.com/apache/pulsar/pull/18712.

### Modifications

* Fix an incorrect reference to trusted certificates.

(cherry picked from commit 4cd64d8ea184e500a1e637510aa54c4c8ac23e96)