[improve] Transactions: use Epoll for eventLoopGroup when available (#1620)
### Motivation

We are using `NioEventLoop` instead of `Epoll`. It is recommended to use
`Epoll`: https://netty.io/wiki/native-transports.html.

### Modifications

* Use the Apache Pulsar method `EventLoopUtil#newEventLoopGroup` to load
the best event loop. Instead of adding any configuration, I am simply
using the defaults that we've already relied on. Please let me know if
we want to make it configurable. Note that 0 threads defaults to 2 times
the number of available processors.

### Verifying this change

This change is trivial, and since the broker already uses this method to
create the event loop, it should be low risk. The only down side is that
Apache Pulsar could move this method and break our usage.

### Documentation

- [x] `no-need-doc`

This is an internal improvement.

(cherry picked from commit 4bc9e4a7a6f1cb08f608f7209055cf4d4b72f08b)